Restructure app files
