+++ app/product.rb
