+++ shopiby.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require "./cart_app"
+require "./app/cart_app"
 
 CartApp.new(ARGV).call
